Was unable to add hycov after the fact.
I got it to work for a moment, just a moment and should have committed
that. I don't think it would have worked in the end though. If I want to
use hycov, I'd have to switch to an older flake, which I can't do after
the fact. Oh well.