Updated NVidia and kernel drivers, and Firefox.
By switching to the latest versions of each of these three things, the
explicit sync crashing seems to have fixed itself. Unfortunately, this
has not fixed the flickering issues in Affinity under Xwayland WINE, and
switching to native Wayland also crashes it. My only option is to use a
VM.