Fixed vconsole setup error messages.
Did this using a workaround, setting the tty to be configured using
xserver settings. However, the tty still hangs, because it turns out
that I've been having a graphics driver issue this whole time. Was not
able to fix this error yet, due to documentation for these kinds of
errors being sparse.