Fixed tty not displaying issue and moved to GRUB.
After trying many many things, the issue is fixed by manually setting
the console resolution using the kernel parameter "video=1920x1080". I
tried doing this first with grub, then with the "vga=", but neither
worked. It's not entirely clear why this issue occurs, but my best guess
is it has something to do with my two monitors having different
resolutions and modesetting and the nvidia framebuffer not being able to
display the higher resolution on the lower resolution monitor.