Added Thunar archive and volume management support.
This was done by adding gvfs in NixOS and polkit-gnome in Home Manager
to allow mounting and authentication for in. For archive management, I
just added file roller to Home Manager since the Thunar plugin does not
come with a archive manager. I also added tumbler, to allow thumbnails
to be displayed.