Moved VM images to new subvolume and added cryptsetup.
It made sense to move the VM images since those can be used regardless
of what OS the VM is on. VMs may be dm-crypt or LUKS encrypted using
cryptsetup, but this has to be managed using an external script until I
find a clean way to explicitly configure and alias this script.