Enabled explicit sync, cleaned comments.
Some comments such as the "see X for more" are overly vague, made those
more specific. Also removed commented-out options I have not touched in
a long time or are irrelevant to the current version of Hyprland. Also
moved misc and render options to the "OTHER" section in hyprland.nix.