Refactored Hyprland keybinds
Now, each group of binds are stored in separate files. The binds
themselves are stored as named lists, allowing different groups of binds
to be easily added together and modified to create a customized Hyprland
keybinding. Also added a default binds option, which keeps the same
binds as the old keybinds.nix did.