Added Hyprspace
Hyprspace is an alternative overview plugin showing all workspaces at
the top of the screen. While I'd prefer if they were visible from the
center, this overview may be more useful since it is workspace-focused
versus window-focused. Hyprspace is accessible using the submap keybind
(enabling simultaneously the submap and Hyprspace).