Added unchartedScripts package with resize-root
The purpose of unchartedScripts is to contain all of my personal scripts
in a single package living in the Nix store, allowing me to call them by
name from the shell. Added the resize-root script, and alias, allowing
me to resize the root tmpfs on the fly. The initial root tmpfs size has
been shrunk to 500M now that resizing it is so simple.