Added nix find impermanent script and alias
This lists every single file in the root tmpfs, aside from symlinks and
files within a ".cache" or "tmp" directory. It also takes an argument
which may specific additional directories to exclude. This is useful for
finding what files on the system are impermanent when setting up
persistence for new programs.