Fixed as many warnings as possible
Includes option name changes related to Firefox and catpuccin. Also
added the isStandalone option to Home Manager to break warnings related
to adding overlays with when useGlobalPkgs is true. Toggling
isStandalone off sets the overlays to null, removing this error.
Restructured the overlays folder to accomodate this, including deleting
the unused Hyprland overlay.