Added uglyAFMode to Hyprland
uglyAFMode turns off shadows, blur, and animations. Useful for saving
battery or on low-end hardware, but it looks ugly AF. This can be
toggled on or off with (SUPER) W. The keybind is made using a new helper
for togglable options associated with variables. uglyAFMode can also be
enabled or disabled by default with a custom option.