Created ethanol and added Affinity bottle
Ethanol is a wine bottle and prefix management script with Home Manager
integration. It serves a similar purpose to the GUI app "Bottles". I
chose to write my own script because 1. Bottles sucks, and 2. Other
nix-based scripts were needlessly complex. The Affinity bottle serves
both as a proof-of-concept for ethanol and as a useful addition to my
configuration.