Added syncthing
Syncthing is a peer-to-peer file synchronization program. This allows me
to automatically share and sync files between devices without having to
upload them to someone else's computer. Note that a new group has been
added and ACLs have been enabled on the HDD to allow all users on
unfathomable-main to read and write to the synchronized directory.