Removed inputs specialArg and extraSpecialArg
This may seem contradictory since my previous commit describes how using
the inputs argument reduces boilerplate, but this is a different
context. Using the inputs argument througout the configuration is
ambiguous, as it doesn't specify which inputs are actually being used by
which modules. But since flakeHelper is essentially an extension of
flake.nix, it is assummed to use every input.