Removed pkgs dependency in myLib
Now when calling enablePkgWith or related functions, a package must be
passed instead of a string (ie pkgs.vim instead of "vim"). The point of
this change is to remove the dependence on a single version of pkgs,
make the functions easier to read, and avoid string interpolation which
was not properly parsing certain packages. The only reason it was done
the way it was previously was to avoid repeatedly writing pkgs, which is
now done using the with keyword.