Moved custom options to Apeiron namespace
Apeiron is a Greek word meaining boundless, indefinite, on the theme of
the rest of the custom utility names in this flake. The main purpose for
moving everything to a new namespace is to create distinction between my
options and other people's options, in a cleaner way than appending
"-home". This could also pave the way for more grouping of related
options. Note that functionally, the configuration is identical.