Replaced enablePkgsWith with mkPkgsModule
The enablePkgsWith function was flawed because it did not support nested
option names and required the option to be configured separately,
reducing only half the boilerplate of making options for each package.
mkPkgsModule solves both these problems by using a list of strings to
access nested options and returning a module declaring both the option
and config.