Sleep now turns off RGB
Saves the current settings on sleep to allow swapping between multiple
colorschemes without having to constantly reenable it. It also now
freezes instead of kills Discord and Slack, suppressing notifications
without needing to reopen the apps on wake. In order to do this, the app
not responding dialog had to get disabled since that causes hypridle to
reawaken.