Enabled zram
Zram replaces swap. Instead of writing to a swapfile on the SSD, it
compresses memory when necessary. Because it doesn't rely on the SSD,
zram is usally significantly faster than swap. It also has the benefit
of reducing SSD writes. Note that swap wasn't being used at all before,
due to fear of it reducing the lifespan of SSDs.