Removed git errors from rebuilds
The git intregration in flakes is actively harmful and cannot be
disabled. As a workaround, the rebuild scripts now copy every file
except for git files to a build directory outside the git repository
before rebuilding. This works because git integration is only enabled if
the flake resides in a git repository.