Added optnix configuration
Scopes for both NixOS and Home Manager have both been added, allowing
searching of options from this flake's NixOS and Home Manager options.
Note the current implementation is broken when using Home Manager
outside NixOS since it adds the scope for nixos despite that scope being
useless. Option evaluation is also broken, but that appears to be a
problem with upstream and not my configuration.