Added master layout
The master layout dedicates most of the screen to one "master" window
and adds the rest to a stack with fixed size. This is helpful for
multitasking because rather than splitting the screen into multiple
awkwardly-sized segments, it dedicates most of the screen to one main
window while keeping other windows accessible. Extra keybinds to switch
layouts and cycle windows help with this.