Started universal keybind management system
The goal is to eventually express all keybinds in a single configuration
format, allowing the same keybinds to automatically be ported to
multiple DEs/WMs. This initial implementation only handles Hyprland and
XFCE and only works for app launcher binds. There is likely also a more
readable way to implement it, so it is subject to change.