feat: add waybar mango integration
The hyprland workspaces module is being copied and converted into the
mango tags module. Also changed hyprland branding to mango branding and
gave the border a background to better match the mango environment. Note
that all of the scratchpad toggles are currently broken since the cli
for sending mango dispatchers is broken.