fix: use string interpolation for the FLAKE variable
Writing references to environment variables directly into a config
caused Home Manager to fail to activate without failing to build! This
is because the activation script runs before environment variables are
defined. To prevent this happening, all references to the FLAKE
environment variable are now substituted using string interpolation
(except for in the rebuild aliases).